// node-sdk/src/openapi/models/ListPayoutWalletsResponse.ts
/**
 * Pluto
 * The API for Pluto
 *
 * OpenAPI spec version: 1.0
 * Contact: support@pluto.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { PayoutWallet } from './PayoutWallet';
import { HttpFile } from '../http/http';

export class ListPayoutWalletsResponse {
    'data'?: Array<PayoutWallet>;
    'has_more'?: boolean;
    'page'?: number;
    'total'?: number;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "data",
            "baseName": "data",
            "type": "Array<PayoutWallet>",
            "format": ""
        },
        {
            "name": "has_more",
            "baseName": "has_more",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "page",
            "baseName": "page",
            "type": "number",
            "format": ""
        },
        {
            "name": "total",
            "baseName": "total",
            "type": "number",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ListPayoutWalletsResponse.attributeTypeMap;
    }

    public constructor() {
    }
}